Refactor tasks into sync and update modules
